requirements.txt
PyQt5>=5.15
numpy>=1.21
Pillow>=9.0